extract in helpers and rename some functions
